Add PragmataPro to .gitignore
PragmataPro is a commercial/proprietary font. I forget how the version
in this repo was obtained, but it sure isn't complete or up-to-date.

Adding it to my .gitignore allows me to keep the incomplete version in
this repo while using the latest version locally, without polluting the
output of `git status`.